+++ Mapsui/Widgets/IWidget.cs
@@ -4,8 +4,8 @@ public interface IWidget
 {
     HorizontalAlignment HorizontalAlignment { get; set; }
     VerticalAlignment VerticalAlignment { get; set; }
-    float MarginX { get; set; }
-    float MarginY { get; set; }
+    double MarginX { get; set; }
+    double MarginY { get; set; }
 
     /// <summary>
     /// The hit box of the widget. This needs to be updated from the widget renderer.
